requirements.txt
django
django-cors-headers
django-rest-framework
python-dotenv
requests
Pillow